% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_degauss_template.R
\name{add_degauss_template}
\alias{add_degauss_template}
\title{Add degauss_template.Rmd file that drives package development}
\usage{
add_degauss_template(
  geomarker = ".",
  overwrite = FALSE,
  open = TRUE,
  dev_dir = "dev"
)
}
\arguments{
\item{geomarker}{Path where to save file}

\item{overwrite}{Whether to overwrite existing degauss_template.Rmd file}

\item{open}{Logical. Whether to open file after creation}

\item{dev_dir}{Name of directory for development Rmarkdown files. Default to "dev".}
}
\value{
Create a degauss_template.Rmd file and return its path
}
\description{
Add degauss_template.Rmd file that drives package development
}
\examples{
# Create a new project
tmpdir <- tempdir()
geomarker_example <- file.path(tmpdir, "geomarker_example")
dir.create(geomarker_example)

# Add
add_degauss_template(geomarker = geomarker_example)

# Delete the example
unlink(geomarker_example, recursive = TRUE)
}
