% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_functions.R
\name{address_is_institutional}
\alias{address_is_institutional}
\title{check if address is a known Cincinnati institutional address}
\usage{
address_is_institutional(address)
}
\arguments{
\item{address}{character vector of address text}
}
\value{
logical vector; TRUE when address contains some text
indicating Cincinnati Children's Hospital, Ronald
McDonald House, Jobs and Family Services
for Hamilton and Butler Counties in Ohio, or
Stetson Square
}
\description{
check if address is a known Cincinnati institutional address
}
\details{
Print the function to view the source and the complete
list of addresses considered to be Cincinnati insitutional addresses;
Note that addresses in other cities might be erroneously categorized as institutional
(e.g., "3333 Burnet Ave Syracuse NY 13206")
}
