% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_functions.R
\name{address_is_nonaddress}
\alias{address_is_nonaddress}
\title{check if address text is not actually an address}
\usage{
address_is_nonaddress(address)
}
\arguments{
\item{address}{character vector of address text}
}
\value{
logical vector; TRUE when address text is "verify",
"foreign", "foreign country", "unknown", or blank.
}
\description{
check if address text is not actually an address
}
