% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_functions.R
\name{address_is_po_box}
\alias{address_is_po_box}
\title{check if address is a PO Box}
\usage{
address_is_po_box(address)
}
\arguments{
\item{address}{character vector of address text}
}
\value{
logical vector; TRUE when address text contains some
variation of "PO Box"
}
\description{
check if address is a PO Box
}
