% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_dates}
\alias{check_dates}
\title{check format of dates from DeGAUSS container input file}
\usage{
check_dates(date, allow_missing = FALSE)
}
\arguments{
\item{date}{vector of dates to be checked for formatting}

\item{allow_missing}{logical. defaults to FALSE, resulting in an error if any dates are missing.}
}
\value{
reformatted vector of dates, or an error if dates could not be reformatted
}
\description{
check format of dates from DeGAUSS container input file
}
\details{
ISO formatted dates (e.g., "\%Y-\%m-\%d" or 2021-01-01) will stay the same
U.S. standard slash formatted dates (common to Microsoft Excel; e.g., "\%m/\%d/\%y" or 1/1/21)
will be reformatted to ISO format
Any input not recognized by one of the two above formats will cause an error and
the user will be instructed to manually reformat their dates.
}
\examples{
\dontrun{
date <- c("1/1/21", "1/2/21", "1/3/21")
check_dates(date)
}
}
