% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_column.R
\name{check_for_column}
\alias{check_for_column}
\title{check for specified columns and corresponding column types}
\usage{
check_for_column(d, column_name, column, column_type = NULL)
}
\arguments{
\item{d}{input dataframe}

\item{column_name}{character string defining name of column to be checked}

\item{column}{character vector to be checked (e.g., d$column_name)}

\item{column_type}{(optional) desired column type to be checked for (e.g., 'character')}
}
\value{
if column_name exists in d and is of the correct column_type, nothing is returned.
if column_name does not exist in d, an error is thrown.
if column is not of the correct column_type, a warning is shown.
}
\description{
check for specified columns and corresponding column types
}
\examples{
\dontrun{
d <- tibble::tribble(
      ~'id', ~'value',
      '123', 123,
      '456', 456
    )
check_for_column(d, 'id', d$id, 'double')
check_for_column(d, 'id2', d$id2, 'double')
}
}
