% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ram.R
\name{check_ram}
\alias{check_ram}
\title{check_ram}
\usage{
check_ram(minimum_ram = 4)
}
\arguments{
\item{minimum_ram}{minimum recommended GB of RAM (as numeric value)}
}
\description{
Checks for amount of system RAM and warns a potential
DeGAUSS user if it might be too low
}
