% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_functions.R
\name{clean_address}
\alias{clean_address}
\title{clean address text strings for geocoding}
\usage{
clean_address(address)
}
\arguments{
\item{address}{character vector of address text}
}
\value{
vector of character strings with non-alphanumerics (except dashes,
which are left in for +4 ZIP issues) and excess white space
removed.
}
\description{
clean address text strings for geocoding
}
