% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{core_lib_images}
\alias{core_lib_images}
\title{list the DeGAUSS images in the core library}
\usage{
core_lib_images(geocoder = TRUE)
}
\arguments{
\item{geocoder}{logical; include "geocoder" in core image list?}
}
\value{
names of DeGAUSS images in the core library as a character vector
}
\description{
list the DeGAUSS images in the core library
}
\examples{
core_lib_images()
core_lib_images(geocoder = FALSE)
}
