% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_menu.R
\name{create_degauss_menu_data}
\alias{create_degauss_menu_data}
\title{create data for use in DeGAUSS menu}
\usage{
create_degauss_menu_data(core_lib_env = get_degauss_core_lib_env())
}
\arguments{
\item{core_lib_env}{a data.frame of info about the DeGAUSS core image
library created with \code{get_degauss_core_lib_env()}}
}
\value{
data.frame of information about core images with arguments
separated into names and default values as well as an added example DeGAUSS command
}
\description{
create data for use in DeGAUSS menu
}
\examples{
dht:::create_degauss_menu_data()
}
