% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_colors.R
\name{degauss_colors}
\alias{degauss_colors}
\title{Access the DeGAUSS color palette}
\usage{
degauss_colors(n)
}
\arguments{
\item{n}{which DeGAUSS color(s): 1-darkblue, 2-lightblue, 3-pink, 4-lightgrey, 5-purple, 6-teal)}
}
\value{
a named character string or vector of named character strings containing RGB colors in hexadecimal
}
\description{
Access the DeGAUSS color palette
}
\examples{
degauss_colors(2)
degauss_colors(1:4)
plot(1:6, rep(1, 6), col = degauss_colors(1:6), pch = 19, cex = 10)
}
