% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_menu.R
\name{degauss_menu}
\alias{degauss_menu}
\title{DeGAUSS Menu}
\usage{
degauss_menu()
}
\description{
Run an interactive shiny application to find geomarkers
available within DeGAUSS based on categories and input data
characteristics. At launch, it will
download the latest information about DeGAUSS images in
the core library.  Suggested DeGAUSS commands are
automatically created and displayed for use.
}
