% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_degauss_files.R
\name{detect_r_version}
\alias{detect_r_version}
\title{Detect current version of R and prompt to update to at least 4.0.0}
\usage{
detect_r_version()
}
\value{
Return current version of R
}
\description{
Detect current version of R and prompt to update to at least 4.0.0
}
