% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dates.R
\name{expand_dates}
\alias{expand_dates}
\title{expand dates between start_date and end_date}
\usage{
expand_dates(d, by)
}
\arguments{
\item{d}{data.frame or tibble with columns called 'start_date' and 'end_date'}

\item{by}{time interval to expand dates (e.g., 'day', 'week', etc)}
}
\value{
long data.frame or tibble with column called 'date' including all dates
between start_date and end_date
}
\description{
expand dates between start_date and end_date
}
\examples{
\dontrun{
d <- data.frame(
  start_date = check_dates(c("1/1/21", "1/2/21", "1/3/21")),
  end_date = check_dates(c("1/7/21", "1/8/21", "1/9/21"))
)
expand_dates(d, by = "day")
}
}
