% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_run.R
\name{find_docker}
\alias{find_docker}
\title{find the path to the docker executable}
\usage{
find_docker()
}
\value{
path to Docker executable found using \code{Sys.which("docker")}
}
\description{
Error if docker cannot be found or if the docker daemon is not running in the background.
}
