% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{get_degauss_core_lib_env}
\alias{get_degauss_core_lib_env}
\title{get DeGAUSS metadata on all images in the \href{https://degauss.org/available_images}{core library}}
\usage{
get_degauss_core_lib_env(..., badges = FALSE)
}
\arguments{
\item{...}{further arguments passed to \code{core_lib_images()}}

\item{badges}{return markdown code for DeGAUSS version and automated build badges?}
}
\value{
data.frame of DeGAUSS metatdata
}
\description{
get DeGAUSS metadata on all images in the \href{https://degauss.org/available_images}{core library}
}
\examples{
get_degauss_core_lib_env(badges = TRUE)

}
