% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeting.R
\name{greeting}
\alias{greeting}
\title{display DeGUASS greeting message in console}
\usage{
greeting(geomarker_name, version, description)
}
\arguments{
\item{geomarker_name}{name of the geomarker, must be the name used in the degauss.org url}

\item{version}{container version number as a character string}

\item{description}{brief description of the container; finishes the sentence "This container..."}
}
\description{
display DeGUASS greeting message in console
}
\details{
greeting message includes name, version, and brief description of container,
as well as links to more information about the specific geomarker,
DeGAUSS in general for troubleshooting, and the DeGUASS user RedCap survey.
}
\examples{
\dontrun{
greeting('roads', '0.4', 'returns proximity and length of nearby major roadways')
}
}
