% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate.R
\name{inflate}
\alias{inflate}
\title{Inflate Rmd to degauss files}
\usage{
inflate(
  geomarker = ".",
  rmd = file.path("dev", "degauss_template.Rmd"),
  test = FALSE,
  overwrite = c("ask", "yes", "no")
)
}
\arguments{
\item{geomarker}{Path to container}

\item{rmd}{Path to Rmarkdown file to inflate}

\item{test}{Logical. Whether to test container after Rmd inflating}

\item{overwrite}{Logical. Whether to overwrite vignette and functions if already exists.}
}
\value{
Return path to current container
}
\description{
Inflate Rmd to degauss files
}
