% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_template.R
\name{inflate_degauss_template}
\alias{inflate_degauss_template}
\title{Inflate Rmd to degauss files}
\usage{
inflate_degauss_template(geomarker = getwd())
}
\arguments{
\item{geomarker}{path to folder containing degauss.Rmd}
}
\value{
Return path to current container
}
\description{
Inflate Rmd to degauss files
}
