% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_menu.R
\name{make_degauss_command}
\alias{make_degauss_command}
\title{creates a \href{https://degauss.org/using_degauss.html#DeGAUSS_Commands}{DeGAUSS command}
with the supplied DeGAUSS image name, version, input file name, and optional argument.}
\usage{
make_degauss_command(
  input_file = "my_address_file_geocoded.csv",
  image,
  version,
  argument = NA
)
}
\arguments{
\item{input_file}{name of input file}

\item{image}{name of DeGAUSS image}

\item{version}{version of DeGAUSS image}

\item{argument}{optional argument}
}
\value{
DeGAUSS command as a character string
}
\description{
creates a \href{https://degauss.org/using_degauss.html#DeGAUSS_Commands}{DeGAUSS command}
with the supplied DeGAUSS image name, version, input file name, and optional argument.
}
\examples{
make_degauss_command(image = "geocoder", version = "3.2.0")
make_degauss_command(image = "geocoder", version = "3.2.0", argument = "0.4")
}
