% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlibrary.R
\name{qlibrary}
\alias{qlibrary}
\title{wrapper for base::library() that automatically supresses package startup messages}
\usage{
qlibrary(...)
}
\arguments{
\item{...}{arguments passed to base::library()}
}
\description{
note that renv will not pickup dependencies loaded using this function
and it is recommended to use something like
\code{withr::with_message_sink("/dev/null", library(dplyr))
} instead
}
\examples{
\dontrun{
qlibrary(dplyr)
}
}
