% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lat_lon_csv.R
\name{read_lat_lon_csv}
\alias{read_lat_lon_csv}
\title{read in and format input file for DeGAUSS container}
\usage{
read_lat_lon_csv(
  filename,
  nest_df = FALSE,
  sf_out = FALSE,
  project_to_crs = NULL
)
}
\arguments{
\item{filename}{name of input file, probably opt$filename if inside container}

\item{nest_df}{logical. If TRUE, data is nested on lat/lon. Defaults to FALSE.}

\item{sf_out}{logical. If TRUE, data is converted as an sf object. Defaults to FALSE.}

\item{project_to_crs}{(optional) if sf_out=TRUE, the crs to which input data is projected. If unspecified
and sf_out=TRUE, the crs defaults to 4326.}
}
\value{
a list with two elements. The first is the raw_data as it is read in from the input file.
The second is a tibble nested on lat and lon to prevent duplication of geomarker computations.
If sf_out=TRUE the second is an sf object.
}
\description{
read in and format input file for DeGAUSS container
}
\examples{
\dontrun{
d <- read_lat_lon_csv(filename = "test/my_address_file_geocoded.csv")
d <- read_lat_lon_csv(
  filename = "test/my_address_file_geocoded.csv",
  sf_out = TRUE, project_to_crs = 5072
)
}

}
