% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_degauss_files.R
\name{use_degauss_compose}
\alias{use_degauss_compose}
\title{use DeGAUSS compose file}
\usage{
use_degauss_compose(...)
}
\arguments{
\item{...}{arguments passed to render_degauss_template (overwrite)}
}
\description{
creates a docker-compose yaml file in current working directory
}
