% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_degauss_files.R
\name{use_degauss_container}
\alias{use_degauss_container}
\title{Use DeGAUSS Container Template}
\usage{
use_degauss_container(geomarker = getwd(), version = "0.1", ...)
}
\arguments{
\item{geomarker}{Path to folder where DeGAUSS container files are to be added;}

\item{version}{string of version number used in freshly created README and entrypoint.R; defaults to "0.1"}

\item{...}{arguments passed to render_degauss_template (overwrite)
defaults to the current working directory}
}
\description{
Creates all the necessary files to create a DeGAUSS container.
The container/geomarker name is assumed to be the basename of the working directory
and the version of R and renv is taken from the calling environment.
This function calls all of the individual \verb{dht::use_degauss_*()} functions to create the following:
\itemize{
\item \code{Dockerfile}
\item \code{Makefile}
\item \code{README.md}
\item \code{entrypoint.R}
\item \code{.dockerignore}
\item \code{.gitignore}
\item \code{test/my_address_file_geocoded.csv}
\item \code{LICENSE.md} GPL license
\item \code{.github/workflows/build-deply.yaml}
}
}
