% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_template.R
\name{use_degauss_template}
\alias{use_degauss_template}
\title{Add degauss.Rmd and Makefile for DeGAUSS container development}
\usage{
use_degauss_template(geomarker = ".")
}
\arguments{
\item{geomarker}{Path where to save file}
}
\value{
Create a degauss_template.Rmd file and return its path
}
\description{
Add degauss.Rmd and Makefile for DeGAUSS container development
}
\examples{
\dontrun{
add_degauss_template(geomarker = geomarker_example)
unlink(geomarker_example, recursive = TRUE)
}
}
