% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geomarker_file.R
\name{write_geomarker_file}
\alias{write_geomarker_file}
\title{write geomarker output to file}
\usage{
write_geomarker_file(
  d,
  raw_data = NULL,
  filename,
  geomarker_name = Sys.getenv("degauss_name"),
  version = Sys.getenv("degauss_version"),
  argument = NULL
)
}
\arguments{
\item{d}{input nest on .row with added geomarker column(s)}

\item{raw_data}{original unnested input data, defaults to NULL (for use when
nest_df = FALSE in read_lat_lon_csv)}

\item{filename}{name of input file, probably opt$filename if inside container}

\item{geomarker_name}{name of the geomarker; defaults to degauss environment variable \code{degauss_name}}

\item{version}{container version number as a character string; defaults to degauss environment
variable \code{degauss_version}}

\item{argument}{optional information to append after the image version number that was specified using a degauss argument;
for example, a selected buffer radius, care site, or geocoding threshold}
}
\value{
output file is written to working directory
}
\description{
write geomarker output to file
}
\examples{
\dontrun{
write_geomarker_file(d$d, d$raw_data,
  filename = "test/my_address_file_geocoded.csv",
  geomarker = "roads", version = "0.4"
)
}

}
