% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_functions.R
\name{address_is_institutional}
\alias{address_is_institutional}
\title{check if address is a known Cincinnati institutional address}
\usage{
address_is_institutional(address)
}
\arguments{
\item{address}{character vector of address text}
}
\value{
logical vector; TRUE when address contains some text
indicating Cincinnati Children's Hospital, Ronald
McDonald House, or Cincinnati Jobs and Family Services.
}
\description{
check if address is a known Cincinnati institutional address
}
