% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{create_degauss_menu_data}
\alias{create_degauss_menu_data}
\title{create data for use in DeGAUSS menu}
\usage{
create_degauss_menu_data(core_images_info = get_core_images())
}
\arguments{
\item{core_images_info}{a data.frame of info about the DeGAUSS core image
library created with \code{get_core_images()}}
}
\value{
data.frame of information about core images with arguments
separated into names and default values as well as an added example DeGAUSS command
}
\description{
create data for use in DeGAUSS menu
}
\examples{
dht:::create_degauss_menu_data()
}
