% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_menu.R
\name{degauss_menu}
\alias{degauss_menu}
\title{DeGAUSS Menu}
\usage{
degauss_menu()
}
\description{
Run an interactive shiny application to create instructions
for running multiple DeGAUSS containers.  At launch, it will
download the latest information about DeGAUSS images in
the core library.  After selection and setting options,
copy (or save to file) bash code used to run a DeGAUSS pipeline.
}
