% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_run.R
\name{degauss_run}
\alias{degauss_run}
\title{run a DeGAUSS container}
\usage{
degauss_run(.x, image, version = "latest", argument = NA, quiet = FALSE)
}
\arguments{
\item{.x}{a data.frame or tibble to be input to a DeGAUSS container}

\item{image}{name of DeGAUSS image}

\item{version}{version of DeGAUSS image; will use latest version if not specified}

\item{argument}{optional argument}

\item{quiet}{suppress output from DeGAUSS container?}
}
\value{
\code{.x} with additional returned DeGAUSS columns
}
\description{
run a DeGAUSS container
}
