% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{get_degauss_env_dockerfile}
\alias{get_degauss_env_dockerfile}
\alias{get_degauss_env_online}
\alias{get_core_images}
\title{get environment variables from a Dockerfile}
\usage{
get_degauss_env_dockerfile(
  dockerfile_path = fs::path_join(c(getwd(), "Dockerfile"))
)

get_degauss_env_online(name = "fortunes")

get_core_images(badges = FALSE)
}
\arguments{
\item{dockerfile_path}{path to Dockerfile}

\item{name}{name of DeGAUSS container to download Dockerfile from}

\item{badges}{return markdown code for DeGAUSS version and automated build?}
}
\value{
named vector of DeGAUSS metatdata elements
}
\description{
These functions look in a Dockerfile to extract environment variables corresponding to DeGAUSS container metadata.
They look for ENV instructions where the name begins with "degauss_";
for example "degauss_name", or "degauss_version".
\code{get_degauss_env_dockerfile} expects a path to a Dockerfile, but \code{get_degauss_env_online}
takes the name of DeGAUSS container that is used to download the
corresponding Dockerfile (from the master branch of the github repository).
It assumes each \code{ENV} instruction is on its own line and defines only one environment variable.
}
\examples{
\dontrun{
use_degauss_dockerfile(version = "0.1")
get_degauss_env_dockerfile()
get_degauss_env_dockerfile()["degauss_version"]
get_degauss_env_online("fortunes")
get_degauss_env_online("fortunes")["degauss_version"]
}

}
