% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{get_env_from_dockerfile}
\alias{get_env_from_dockerfile}
\alias{get_env_from_degauss_online}
\title{get environment variables from a Dockerfile}
\usage{
get_env_from_dockerfile(
  dockerfile_path = fs::path_join(c(getwd(), "Dockerfile"))
)

get_env_from_degauss_online(name = "fortunes")
}
\arguments{
\item{dockerfile_path}{path to Dockerfile}

\item{name}{name of DeGAUSS container to download Dockerfile from}
}
\value{
named list of DeGAUSS metatdata elements:
"degauss_name", "degauss_version", "degauss_description", "degauss_argument"
}
\description{
These functions look in a Dockerfile to extract specified environment variables and
their values corresponding to DeGAUSS container metadata.
\code{get_env_from_dockerfile} expects a path to a Dockerfile, but \code{get_env_from_degauss_online}
takes the name of DeGAUSS container that is used to download the corresponding Dockerfile.
It assumes each \code{ENV} instruction is on its own line,
defines only one environment variable, and is always in the order:
"degauss_name", "degauss_version", "degauss_description", "degauss_argument"
}
\examples{
\dontrun{
use_degauss_dockerfile(version = "0.1")
get_env_from_dockerfile()
get_env_from_dockerfile()["degauss_version"]
get_env_from_degauss_online("fortunes")
get_env_from_degauss_online("fortunes")["degauss_version"]
}
}
