% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_env.R
\name{get_env_from_dockerfile}
\alias{get_env_from_dockerfile}
\title{get environment variables from a Dockerfile}
\usage{
get_env_from_dockerfile(
  dockerfile_path = fs::path_join(c(getwd(), "Dockerfile")),
  env_names = c("degauss_name", "degauss_version", "degauss_description")
)
}
\arguments{
\item{dockerfile_path}{path to Dockerfile}

\item{env_names}{names of environment variables to extract}
}
\description{
This function looks in a Dockerfile to extract specified environment variables and their values.
It assumes each \code{ENV} instruction is on its own line and defines only one environment variable.
The default names of environment variables assumes those used for DeGAUSS metadata.
}
\examples{
\dontrun{
use_degauss_dockerfile(version = "0.1")
get_env_from_dockerfile()
get_env_from_dockerfile()["degauss_version"]
}
}
