% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeting.R
\name{greeting}
\alias{greeting}
\title{display DeGUASS greeting message in console}
\usage{
greeting(
  geomarker_name = Sys.getenv("degauss_name"),
  version = Sys.getenv("degauss_version"),
  description = Sys.getenv("degauss_description")
)
}
\arguments{
\item{geomarker_name}{name of the geomarker, must be the name used in the degauss.org url}

\item{version}{container version number as a character string}

\item{description}{brief description of the container; finishes the sentence "This container returns..."}
}
\description{
if not supplied as arguments, greeting-specific values
(geomarker_name, version, description) are read in from the environment variables
specified in the Dockerfile and made available when running the container;
these include \code{degauss_name}, \code{degauss_version}, and \code{degauss_description}
}
\details{
greeting message includes name, version, and brief description of container,
as well as a link to more information about the specific geomarker
}
\examples{
\dontrun{
greeting('roads', '0.4', 'returns proximity and length of nearby major roadways')
}
}
