% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degauss_run.R
\name{make_degauss_command}
\alias{make_degauss_command}
\title{create a \href{https://degauss.org/using_degauss.html#DeGAUSS_Commands}{DeGAUSS command}}
\usage{
make_degauss_command(
  input_file = "my_address_file_geocoded.csv",
  image,
  version = "latest",
  argument = NA,
  docker_cmd = "docker"
)
}
\arguments{
\item{input_file}{name of input file}

\item{image}{name of DeGAUSS image}

\item{version}{version of DeGAUSS image}

\item{argument}{optional argument}

\item{docker_cmd}{path to docker executable}
}
\value{
DeGAUSS command as a character string
}
\description{
create a \href{https://degauss.org/using_degauss.html#DeGAUSS_Commands}{DeGAUSS command}
}
\examples{
make_degauss_command(image = "geocoder", version = "3.2.0")
make_degauss_command(image = "geocoder", version = "3.2.0", argument = "0.4")
make_degauss_command(image = "geocoder", version = "3.2.0", docker_cmd = "/usr/local/bin/docker")
}
