% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlibrary.R
\name{qlibrary}
\alias{qlibrary}
\title{wrapper for base::library() that automatically supresses package startup messages}
\usage{
qlibrary(...)
}
\arguments{
\item{...}{arguments passed to base::library()}
}
\description{
wrapper for base::library() that automatically supresses package startup messages
}
\examples{
\dontrun{
qlibrary(dplyr)
}
}
