% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_degauss_files.R
\name{use_degauss_container}
\alias{use_degauss_container}
\title{Use DeGAUSS Container Template}
\usage{
use_degauss_container(geomarker = getwd(), ...)
}
\arguments{
\item{geomarker}{Path to folder where DeGAUSS container files are to be added;}

\item{...}{arguments passed to render_degauss_template
defaults to the current working directory}
}
\description{
This function calls all of the individual \verb{dht::use_degauss_*()} functions to add:
\itemize{
\item \code{Dockerfile}
\item \code{Makefile}
\item \code{README.md}
\item \code{entrypoint.R}
\item \code{.dockerignore}
\item \code{.gitignore}
\item \code{test/my_address_file_geocoded.csv}
\item \code{LICENSE.md} GPL license
\item \code{.github/workflows/build-deply.yaml}
}
}
