% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_degauss_template.R
\name{use_degauss_template}
\alias{use_degauss_template}
\title{Use DeGAUSS Template}
\usage{
use_degauss_template(path = fs::path_wd(), ...)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists. The terminal directory
should be named for the geomarker (e.g., 'home/folder/geomarker_name')}

\item{...}{additional arguments passed to \code{use_degauss_dockerfile} used
to specify the base image and renv version}
}
\value{
Path to the newly created degauss directory, invisibly.
}
\description{
This function creates the following:
\itemize{
\item \code{Dockerfile.txt}
\item \code{Makefile.txt}
\item \code{README.md}
\item \verb{R script}
\item \code{.dockerignore}
\item \code{.gitignore}
\item \code{test/my_address_file_geocoded.csv}
\item \code{LICENSE.md} GPL license
\item renv infrastructure
}
}
