% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_geomarker_file.R
\name{write_geomarker_file}
\alias{write_geomarker_file}
\title{write geomarker output to file}
\usage{
write_geomarker_file(d, raw_data, filename, geomarker_name, version)
}
\arguments{
\item{d}{input nest on .row with added geomarker column(s)}

\item{raw_data}{original unnested input data}

\item{filename}{name of input file, probably opt$filename if inside container}

\item{geomarker_name}{name of the geomarker, must be the name used in the degauss.org url}

\item{version}{container version number as a character string}
}
\value{
output file is written to working directory
}
\description{
write geomarker output to file
}
\examples{
\dontrun{
write_geomarker_file(d$d, d$raw_data,
           filename='test/my_address_file_geocoded.csv',
           geomarker='roads', version='0.4')
}
}
